+++ backend-ping-pong/src/routes/duel.ts
@@ -2,7 +2,7 @@ import { FastifyPluginAsync } from "fastify"
 import websocket from '@fastify/websocket'
 import { v4 as uuidv4 } from 'uuid'
 import { MatchManager } from "../duel/MatchManager"
-import { Player } from "../duel/Player"
+import { Player } from "../utils/Player"
 
 const example: FastifyPluginAsync = async (fastify, opts): Promise<void> => {
     await fastify.register(websocket)
