+++ backend-ping-pong/src/utils/Rectangle.ts
